feat: Making a base class for Static Components (#192)
