+++ bundles/core/src/main/java/com/adobe/cq/forms/core/components/models/form/Base.java
@@ -18,22 +18,13 @@ package com.adobe.cq.forms.core.components.models.form;
 import java.util.Collections;
 import java.util.Map;
 
-import javax.annotation.Nonnull;
-
-import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.osgi.annotation.versioning.ConsumerType;
 
 import com.adobe.cq.export.json.ComponentExporter;
-import com.adobe.cq.forms.core.components.views.Views;
-import com.adobe.cq.wcm.core.components.models.Component;
-import com.day.cq.i18n.I18n;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import com.fasterxml.jackson.annotation.JsonValue;
-import com.fasterxml.jackson.annotation.JsonView;
 
 /**
  * A base interface to be extended by all the different types of form elements.
@@ -41,197 +32,8 @@ import com.fasterxml.jackson.annotation.JsonView;
  * @since com.adobe.cq.forms.core.components.models.form 0.0.1
  */
 @ConsumerType
-public interface Base extends Component {
-
-    public final String CUSTOM_PROPERTY_WRAPPER = "af:layout";
+public interface Base extends FormComponent {
     public final String DATE_FORMATTER = "yyyy-MM-dd";
-
-    /**
-     * Defines the view type. Possible values: {@code text-input}, {@code multiline-input}, {@code number-input}, {@code date-input},
-     * {@code file-input}, {@code drop-down}, {@code radio-group}
-     * , {@code plain-text}, {@code checkbox}, {@code button}, {@code panel}
-     *
-     * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-     */
-    public enum FieldType {
-        TEXT_INPUT("text-input"),
-        MULTILINE_INPUT("multiline-input"),
-        NUMBER_INPUT("number-input"),
-        DATE_INPUT("date-input"),
-        FILE_INPUT("file-input"),
-        DROP_DOWN("drop-down"),
-        RADIO_GROUP("radio-group"),
-        PLAIN_TEXT("plain-text"),
-        CHECKBOX("checkbox"),
-        BUTTON("button"),
-        PANEL("panel"),
-        FORM("form"),
-        CHECKBOX_GROUP("checkbox-group"),
-        TEXT("text");
-
-        private String value;
-
-        FieldType(String value) {
-            this.value = value;
-        }
-
-        /**
-         * Given a {@link String} <code>value</code>, this method returns the enum's value that corresponds to the provided string
-         * representation. If no representation is found, {@link #TEXT_INPUT} will be returned.
-         *
-         * @param value the string representation for which an enum value should be returned
-         * @return the corresponding enum value, if one was found, or {@link #TEXT_INPUT}
-         * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-         */
-        public static FieldType fromString(String value) {
-            for (FieldType type : FieldType.values()) {
-                if (StringUtils.equals(value, type.value)) {
-                    return type;
-                }
-            }
-            return TEXT_INPUT;
-        }
-
-        /**
-         * Returns the string value of this enum constant.
-         *
-         * @return the string value of this enum constant
-         * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-         */
-        public String getValue() {
-            return value;
-        }
-
-        @Override
-        @JsonValue
-        public String toString() {
-            return value;
-        }
-    }
-
-    /**
-     * Defines the constraint type. Possible values: {@code type}, {@code required}, {@code minimum}, {@code maximum}, {@code minLength},
-     * {@code maxLength}
-     * , {@code step}, {@code format}, {@code pattern}, {@code minItems}, {@code maxItems}, {@code uniqueItems}, {@code enforceEnum},
-     * {@code validationExpression}
-     *
-     * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-     */
-    public enum ConstraintType {
-        TYPE("type"),
-        REQUIRED("required"),
-        MINIMUM("minimum"),
-        MAXIMUM("maximum"),
-        MIN_LENGTH("minLength"),
-        MAX_LENGTH("maxLength"),
-        STEP("step"),
-        FORMAT("format"),
-        PATTERN("pattern"),
-        MIN_ITEMS("minItems"),
-        MAX_ITEMS("maxItems"),
-        UNIQUE_ITEMS("uniqueItems"),
-        ENFORCE_ENUM("enforceEnum"),
-        VALIDATION_EXPRESSION("validationExpression");
-
-        private String value;
-
-        ConstraintType(String value) {
-            this.value = value;
-        }
-
-        /**
-         * Given a {@link String} <code>value</code>, this method returns the enum's value that corresponds to the provided string
-         * representation
-         *
-         * @param value the string representation for which an enum value should be returned
-         * @return the corresponding enum value, if one was found
-         * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-         */
-        public static ConstraintType fromString(String value) {
-            for (ConstraintType type : ConstraintType.values()) {
-                if (StringUtils.equals(value, type.value)) {
-                    return type;
-                }
-            }
-            return null;
-        }
-
-        /**
-         * Returns the string value of this enum constant.
-         *
-         * @return the string value of this enum constant
-         * @since com.adobe.cq.wcm.core.components.models.form 13.0.0
-         */
-        public String getValue() {
-            return value;
-        }
-
-        @Override
-        @JsonValue
-        public String toString() {
-            return value;
-        }
-
-        /**
-         * returns the name of the property which stores the value of the error message for the constraint
-         * 
-         * @return string
-         */
-        public String getMessageProperty() {
-            return value + "Message";
-        }
-    }
-
-    /**
-     * Defines the assist priority type. Possible values: {@code custom}, {@code description}, {@code label}, {@code name}
-     *
-     * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-     */
-    public enum AssistPriority {
-        CUSTOM("custom"),
-        DESCRIPTION("description"),
-        LABEL("label"),
-        NAME("name");
-
-        private String value;
-
-        AssistPriority(String value) {
-            this.value = value;
-        }
-
-        /**
-         * Given a {@link String} <code>value</code>, this method returns the enum's value that corresponds to the provided string
-         * representation
-         *
-         * @param value the string representation for which an enum value should be returned
-         * @return the corresponding enum value, if one was found
-         * @since com.adobe.cq.wcm.core.components.models.form 13.0.0
-         */
-        public static AssistPriority fromString(String value) {
-            for (AssistPriority type : AssistPriority.values()) {
-                if (StringUtils.equals(value, type.value)) {
-                    return type;
-                }
-            }
-            return null;
-        }
-
-        /**
-         * Returns the string value of this enum constant.
-         *
-         * @return the string value of this enum constant
-         * @since com.adobe.cq.wcm.core.components.models.form 13.0.0
-         */
-        public String getValue() {
-            return value;
-        }
-
-        @Override
-        @JsonValue
-        public String toString() {
-            return value;
-        }
-    }
 
     /**
      * Returns label of the form field
@@ -241,28 +43,6 @@ public interface Base extends Component {
      */
     @Nullable
     default Label getLabel() {
-        return null;
-    }
-
-    /**
-     * Returns the name of the form field
-     *
-     * @return name of the form field
-     * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-     */
-    @Nullable
-    default String getName() {
-        return null;
-    }
-
-    /**
-     * Returns the reference to the data model
-     *
-     * @return reference to the data model
-     * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-     */
-    @Nullable
-    default String getDataRef() {
         return null;
     }
 
@@ -302,24 +82,6 @@ public interface Base extends Component {
     }
 
     /**
-     * Returns the field type
-     *
-     * @return the field type
-     * @since com.adobe.cq.forms.core.components.models.form 1.0.0
-     */
-    String getFieldType(); // todo: keeping string here to support custom view types
-
-    /**
-     * Returns {@code true} if form field should be visible, otherwise {@code false}.
-     *
-     * @return {@code true} if form field should be visible, otherwise {@code false}
-     * @since com.adobe.cq.forms.core.components.models.form 0.0.1
-     */
-    default boolean isVisible() {
-        return true;
-    }
-
-    /**
      * Returns {@code true} if form field should be enabled, otherwise {@code false}.
      *
      * @return {@code true} if form field should be enabled, otherwise {@code false}
@@ -353,29 +115,6 @@ public interface Base extends Component {
     }
 
     /**
-     * Custom Properties
-     * 
-     * @since com.adobe.cq.forms.core.components.models.form 1.1.0
-     */
-    @NotNull
-    @JsonInclude(Include.NON_EMPTY)
-    default Map<String, Object> getProperties() {
-        return Collections.emptyMap();
-    }
-
-    /**
-     * Custom DOR Properties
-     *
-     * @since com.adobe.cq.forms.core.components.models.form 2.1.0
-     */
-
-    @JsonView(Views.Author.class)
-    @JsonInclude(Include.NON_EMPTY)
-    default Map<String, Object> getDorProperties() {
-        return Collections.emptyMap();
-    }
-
-    /**
      * Returns the tool tip of the field
      *
      * @return the tool tip of the field
@@ -395,40 +134,5 @@ public interface Base extends Component {
     @JsonIgnore
     default boolean isTooltipVisible() {
         return false;
-    }
-
-    /**
-     * Returns the rules defined for the component after filtering out invalid rules
-     * If no rules are defined, returns an empty map
-     * 
-     * @return map containing the rules and their expressions
-     */
-    @NotNull
-    @JsonInclude(Include.NON_EMPTY)
-    default Map<String, String> getRules() {
-        return Collections.emptyMap();
-    }
-
-    /**
-     * Returns the events defined for the component after filtering out invalid rules
-     * 
-     * @return map containing the events and their expressions
-     *         If no rules are defined, returns an empty map
-     */
-    @NotNull
-    @JsonInclude(Include.NON_EMPTY)
-    default Map<String, String[]> getEvents() {
-        return Collections.emptyMap();
-    }
-
-    /**
-     * Sets i18n object
-     * 
-     * @param i18n reference to the {@link I18n} object
-     * @since com.adobe.cq.forms.core.components.models.form 2.0.0
-     */
-    @JsonIgnore
-    default void setI18n(@Nonnull I18n i18n) {
-        // empty body
     }
 }
