+++ bundles/core/src/main/java/com/adobe/cq/forms/core/components/util/package-info.java
@@ -18,7 +18,7 @@
  * This package defines utility classes exposed by the Adobe Experience Manager Core Forms Components Bundle.
  * </p>
  */
-@Version("1.2.0")
+@Version("2.0.0")
 package com.adobe.cq.forms.core.components.util;
 
 import org.osgi.annotation.versioning.Version;
