+++ bundles/core/src/test/java/com/adobe/cq/forms/core/components/internal/models/v1/form/CheckBoxGroupImplTest.java
@@ -30,6 +30,8 @@ import com.adobe.cq.forms.core.components.internal.form.FormConstants;
 import com.adobe.cq.forms.core.components.models.form.Base;
 import com.adobe.cq.forms.core.components.models.form.BaseConstraint;
 import com.adobe.cq.forms.core.components.models.form.CheckBoxGroup;
+import com.adobe.cq.forms.core.components.models.form.ConstraintType;
+import com.adobe.cq.forms.core.components.models.form.FieldType;
 import com.adobe.cq.forms.core.components.models.form.Label;
 import com.adobe.cq.forms.core.context.FormsCoreComponentTestContext;
 import com.adobe.cq.wcm.style.ComponentStyleInfo;
@@ -68,7 +70,7 @@ public class CheckBoxGroupImplTest {
     @Test
     void testFieldType() {
         CheckBoxGroup checkboxGroup = getCheckBoxGroupUnderTest(PATH_CHECKBOX_GROUP);
-        assertEquals(Base.FieldType.CHECKBOX_GROUP.getValue(), checkboxGroup.getFieldType());
+        assertEquals(FieldType.CHECKBOX_GROUP.getValue(), checkboxGroup.getFieldType());
     }
 
     @Test
@@ -199,8 +201,8 @@ public class CheckBoxGroupImplTest {
     @Test
     void testGetConstraintMessages() {
         CheckBoxGroup checkboxGroup = getCheckBoxGroupUnderTest(PATH_CHECKBOX_GROUP);
-        Map<Base.ConstraintType, String> constraintsMessages = checkboxGroup.getConstraintMessages();
-        assertEquals(constraintsMessages.get(Base.ConstraintType.TYPE), "incorrect type");
+        Map<ConstraintType, String> constraintsMessages = checkboxGroup.getConstraintMessages();
+        assertEquals(constraintsMessages.get(ConstraintType.TYPE), "incorrect type");
         CheckBoxGroup checkboxGroupMock = Mockito.mock(CheckBoxGroup.class);
         Mockito.when(checkboxGroupMock.getConstraintMessages()).thenCallRealMethod();
         assertEquals(Collections.emptyMap(), checkboxGroupMock.getConstraintMessages());
