+++ bundles/core/src/test/java/com/adobe/cq/forms/core/components/internal/models/v1/form/DatePickerImplTest.java
@@ -61,7 +61,7 @@ public class DatePickerImplTest {
     @Test
     void testFieldType() {
         DatePicker datePicker = Utils.getComponentUnderTest(PATH_DATEPICKER, DatePicker.class, context);
-        assertEquals(Base.FieldType.DATE_INPUT.getValue(), datePicker.getFieldType());
+        assertEquals(FieldType.DATE_INPUT.getValue(), datePicker.getFieldType());
     }
 
     @Test
@@ -195,8 +195,8 @@ public class DatePickerImplTest {
     @Test
     void testGetConstraintMessages() {
         DatePicker datePicker = Utils.getComponentUnderTest(PATH_DATEPICKER, DatePicker.class, context);
-        Map<Base.ConstraintType, String> constraintsMessages = datePicker.getConstraintMessages();
-        assertEquals(constraintsMessages.get(Base.ConstraintType.TYPE), "incorrect type");
+        Map<ConstraintType, String> constraintsMessages = datePicker.getConstraintMessages();
+        assertEquals(constraintsMessages.get(ConstraintType.TYPE), "incorrect type");
         DatePicker datePickerMock = Mockito.mock(DatePicker.class);
         Mockito.when(datePickerMock.getConstraintMessages()).thenCallRealMethod();
         assertEquals(Collections.emptyMap(), datePickerMock.getConstraintMessages());
