+++ bundles/core/src/test/java/com/adobe/cq/forms/core/components/internal/models/v1/form/DropDownImplTest.java
@@ -66,7 +66,7 @@ public class DropDownImplTest {
     @Test
     void testFieldType() {
         DropDown dropdown = Utils.getComponentUnderTest(PATH_DROPDOWN, DropDown.class, context);
-        assertEquals(Base.FieldType.DROP_DOWN.getValue(), dropdown.getFieldType());
+        assertEquals(FieldType.DROP_DOWN.getValue(), dropdown.getFieldType());
     }
 
     @Test
@@ -197,8 +197,8 @@ public class DropDownImplTest {
     @Test
     void testGetConstraintMessages() {
         DropDown dropdown = Utils.getComponentUnderTest(PATH_DROPDOWN, DropDown.class, context);
-        Map<Base.ConstraintType, String> constraintsMessages = dropdown.getConstraintMessages();
-        assertEquals(constraintsMessages.get(Base.ConstraintType.TYPE), "incorrect type");
+        Map<ConstraintType, String> constraintsMessages = dropdown.getConstraintMessages();
+        assertEquals(constraintsMessages.get(ConstraintType.TYPE), "incorrect type");
         DropDown dropdownMock = Mockito.mock(DropDown.class);
         Mockito.when(dropdownMock.getConstraintMessages()).thenCallRealMethod();
         assertEquals(Collections.emptyMap(), dropdownMock.getConstraintMessages());
