+++ bundles/core/src/test/resources/form/checkbox/exporter-checkbox.json
@@ -9,7 +9,7 @@
   "description" : "dummy",
   "tooltip": "test-short-description",
   "properties": {
-    "af:layout" : {
+    "afs:layout" : {
       "tooltipVisible": false,
       "orientation" : "horizontal"
     },
