+++ bundles/core/src/test/resources/form/datepicker/exporter-datepicker-message.json
@@ -9,7 +9,7 @@
   "description" : "dummy",
   "tooltip": "test-short-description",
   "properties": {
-    "af:layout" : {
+    "afs:layout" : {
       "tooltipVisible": false
     },
     "fd:dor": {
