+++ bundles/core/src/test/resources/form/numberinput/exporter-numberinput.json
@@ -9,7 +9,7 @@
   "description" : "dummy",
   "tooltip": "test-short-description",
   "properties": {
-    "af:layout" : {
+    "afs:layout" : {
       "tooltipVisible": false,
       "leadDigits": 2,
       "fracDigits" : 4
