+++ bundles/core/src/test/resources/form/panel/exporter-array-panel.json
@@ -39,7 +39,7 @@
     "value":"dependent names"
   },
   "properties":{
-    "af:layout":{
+    "afs:layout":{
       "tooltipVisible":false
     }
   },
