+++ bundles/core/src/test/resources/form/panel/exporter-panel.json
@@ -37,7 +37,7 @@
     "value":"dependent names"
   },
   "properties":{
-    "af:layout":
+    "afs:layout":
     {
       "tooltipVisible":false
     }
