+++ bundles/core/src/test/resources/form/panel/exporter-rules-panel.json
@@ -42,7 +42,7 @@
     ]
   },
   "properties":{
-    "af:layout":{
+    "afs:layout":{
       "tooltipVisible":false
     }
   },
