+++ bundles/core/src/test/resources/form/panelcontainer/exporter-array-panelcontainer.json
@@ -39,7 +39,7 @@
     "value":"dependent names"
   },
   "properties":{
-    "af:layout":{
+    "afs:layout":{
       "tooltipVisible":false
     }
   },
