+++ bundles/core/src/test/resources/form/panelcontainer/exporter-panelcontainer.json
@@ -37,7 +37,7 @@
     "value":"dependent names"
   },
   "properties": {
-    "af:layout": {
+    "afs:layout": {
       "tooltipVisible":false
     }
   },
