+++ bundles/core/src/test/resources/form/panelcontainer/exporter-rules-panelcontainer.json
@@ -42,7 +42,7 @@
     ]
   },
   "properties":{
-    "af:layout":{
+    "afs:layout":{
       "tooltipVisible":false
     }
   },
