+++ bundles/core/src/test/resources/form/text/test-content.json
@@ -4,11 +4,9 @@
     "sling:resourceType": "core/fd/components/form/text/v1/text",
     "name" : "abc",
     "value" : "<p>xyz</p>",
-    "jcr:title" : "Text",
-    "hideTitle" : false,
     "visible" : true,
     "dataRef" : "a.b",
-    "fieldType": "text",
+    "fieldType": "plain-text",
     "textIsRich": true
   }
 }
