+++ bundles/core/src/test/resources/schema/0.11.0-Pre/adaptive-form.schema.json
@@ -715,7 +715,8 @@
             "screenReaderText",
             "rules",
             "events",
-            "properties"
+            "properties",
+            "id"
           ],
           "errorMessage": {
             "enum": "${0} property is not allowed in StaticField"
